requirements.txt
alembic==1.4.2
apispec==3.3.0
apispec-webframeworks==0.5.2
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
flask-apispec==0.8.8
Flask-Cors==3.0.8
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.5.2
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.2
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
ruamel.appconfig==0.5.4
ruamel.std.argparse==0.8.3
six==1.15.0
SQLAlchemy==1.3.16
SQLAlchemy-Utils==0.36.6
urllib3==1.25.10
wcwidth==0.2.5
webargs==5.5.3
Werkzeug==1.0.1
yaml-1.3==0.1.0
